Started working on handler type alias and supporting functions
