+++ internal/handlers/irc/irc.go
@@ -13,7 +13,8 @@ and the IRCSettings that were passed into NewClient
 */
 type Client struct {
 	*girc.Client
-	Settings internal.IRCSettings
+	Settings   internal.IRCSettings
+	HandlerMap map[string]Handler
 }
 
 /*
@@ -33,7 +34,7 @@ func NewClient(settings internal.IRCSettings) Client {
 			Pass: settings.NickServPassword,
 		}
 	}
-	return Client{client, settings}
+	return Client{client, settings, GetHandlerMapping()}
 }
 
 /*
@@ -64,20 +65,7 @@ that were passed in to NewClient
 */
 func (c Client) addHandlers() {
 	fmt.Println("Adding IRC event handlers...")
-	c.Handlers.Add(girc.PRIVMSG, func(gc *girc.Client, e girc.Event) {
-		if pretty, ok := e.Pretty(); ok {
-			c.SendMessage(pretty)
-		}
-	})
-	c.Handlers.Add(girc.CONNECTED, connectHandler(c))
-}
-
-/*
-connectHandler return a function to use as the connect handler for girc,
-so that the specified channel is joined after the server connection is established
-*/
-func connectHandler(c Client) func(*girc.Client, girc.Event) {
-	return func(gc *girc.Client, e girc.Event) {
-		c.Cmd.Join(c.Settings.Channel)
+	for eventType, handler := range c.HandlerMap {
+		c.Handlers.Add(eventType, handler(c))
 	}
 }
